Main

Documentation

type QueueMapVar = TVar (Map UUID (TQueue ByteString))

getUUID :: ActionM UUID

getUUID parses the uuid parameter - looking in capture, then in form data, then in query parameters - and parses it, failing with HTTP 400 (Bad Request) on parse failure.

getAction :: QueueMapVar -> ActionM ()

getAction takes the next value from a queue and returns it with HTTP 200 (OK), or if the queue is empty or non-existent, returns nothing with HTTP 204 (No Content).

postAction :: QueueMapVar -> ActionM ()

postAction adds a value in the form field value to a queue, creating the queue if necessary, and then returns HTTP 200 (OK).

deleteAction :: QueueMapVar -> ActionM ()

deleteAction deletes a queue, if it exists, and returns HTTP 200 (OK).

main :: IO ()